+++ CHANGELOG.md
@@ -6,7 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## [Unreleased]
 
-Nothing yet!
+Nothing yet.
+
+## [0.16.0] - 2026-03-15
+
+### Changed
+
+- Update `askama` to `0.15`
+- Update `r2d2_sqlite` to `0.32` and use bundled sqlite instead of standalone `rusqlite`
+- Update `sentry` to `0.47`
+- Improve Sentry tracing registration
 
 ## [0.15.0] - 2025-12-07
 
@@ -306,7 +315,8 @@ Nothing yet!
   - Create a "watcher" for a single playlist
   - Background task runs on an interval, transferring tracks for each user that has configured a "watcher"
 
-[Unreleased]: https://github.com/zaknesler/modulate/compare/v0.15.0...HEAD
+[Unreleased]: https://github.com/zaknesler/modulate/compare/v0.16.0...HEAD
+[0.16.0]: https://github.com/zaknesler/modulate/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/zaknesler/modulate/compare/v0.14.4...v0.15.0
 [0.14.4]: https://github.com/zaknesler/modulate/compare/v0.14.3...v0.14.4
 [0.14.3]: https://github.com/zaknesler/modulate/compare/v0.14.2...v0.14.3